Store the datetime field from the hardware file header as a long-long, as opposed to a long.

+++ Source/Common/HardwareModel/P_engine.h
@@ -64,7 +64,7 @@ public:
     /* Engines may be created from a configuration file. If so, some portion of
      * their metadata will be set. */
     std::string author;
-    long datetime;
+    long long datetime;
     std::string version;
     std::string fileOrigin;
